requirements.txt


opencv-python
imageio
imageio-ffmpeg
pytorch-lightning
omegaconf
einops
transformers
matplotlib
pillow
# pip install pytorch-lightning==1.6.5
# pip install scikit-learn
# pip install scipy
# pip install torch==2.1.0 torchvision==0.16.0 torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu121
